Extract behaviors to separate file
